requirements.txt
kivy==2.1.0
https://github.com/kivymd/KivyMD/archive/refs/heads/master.zip
pyjnius==1.4.2
openssl
requests==2.28.2
pillow==9.5.0
